Fix cursor motion in bidi text and implement direction
Each cursor/selection position now has a bias associated with it that
determines whether it is associated with the character before or after
it. This allows proper displaying of the cursor around line breaks,
and improved handling of bidi text.

The bidi algorithm is extended to take line wrapping into account, and
several related pieces of the core are updated to handle the changed
format.

The editor base direction can now be configured, and each line can
have its own direction (ltr or rtl).